[refactor, feat] refactored package layout and added support for Tables in Datastore
